implement camera_eeprom factory and use camera_eeprom interface

+++ tools/eeprom-tool/eeprom_list.h
@@ -6,15 +6,15 @@
 #include <map>
 
 struct EepromProperties{
-    uint16_t size;
-    EepromProperties(uint16_t _size):
+    uint32_t size;
+    EepromProperties(uint32_t _size):
         size(_size)
         {}
 };
 
 const std::map<std::string, EepromProperties> EEPROMS = {
-    {"24c1024", EepromProperties((uint16_t)131072)},
-    {"demo", EepromProperties((uint16_t)42)}
+    {"24c1024", EepromProperties((uint32_t)131072)},
+    {"demo", EepromProperties((uint32_t)42)}
 };
 
 #endif
